Add files via upload
There are 12 inches in a foot and 3 feet in a yard. Create a class named InchConversion. Its main() method 
accepts a value in inches from a user at the keyboard and then passes the entered value to two methods. 
One converts the value from inches to feet, and the other converts the same value from inches to yards. Each 
method displays the results with appropriate explanation. Save the application as InchConversion.java